Fix API to be useable in tests
Previously, Can's modes could only be set via ARGV. This commit
implements a new module function `cli`, that runs the main `can`
function with ARGV as the list of options. Options can be passed to Can
in an application by providing the `can` method with a list of
arguments, e.g. to find info on file `foo`: `Can.can(%w[-n foo])`.